fix(core): break type-module cycle in config types
Commit 60dca9df1 (config-types split) introduced a type-level import
cycle between packages/core/src/types/config/autonomy.ts and
runtime.ts: autonomy.ts imported FleetChatVerbosity from runtime.ts,
while runtime.ts imported AutonomyConfig from autonomy.ts. This raised
the type-inclusive cycle count to 16 vs the 15-entry baseline in
architecture/exceptions.json, failing the CI architecture-health check.

Move FleetChatVerbosity, FLEET_CHAT_VERBOSITY_VALUES, and
resolveFleetChatVerbosity into a new dependency-free leaf module
config/fleet-chat.ts. resolveFleetChatVerbosity now takes a minimal
structural param ({ fleetChatVerbosity? }) instead of
Pick<AutonomyConfig, ...>, so the leaf imports nothing from the config
dir. Both autonomy.ts and runtime.ts import from fleet-chat.ts, and the
config barrel re-exports it, preserving the public surface.

Verified in an isolated worktree off origin/main:
- pnpm --filter @wrongstack/core typecheck: 0 errors
- architecture health: type cycles 16 -> 15, unexcepted 0
